# [WIP] Reuse built-in definitions for CRDs