--- a/staging/src/k8s.io/kms/internal/plugins/_mock/go.mod
+++ b/staging/src/k8s.io/kms/internal/plugins/_mock/go.mod
@@ -25,3 +25,5 @@ replace (
 	k8s.io/apimachinery => ../../../../apimachinery
 	k8s.io/kms => ../../../../kms
 )
+
+replace k8s.io/kube-openapi => github.com/jefftree/kube-openapi v0.0.8-gnostic.0.20231025212924-d8bffb57e175