<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide/first-contribution.md#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

#### What type of PR is this?

/kind cleanup

#### What this PR does / why we need it:

The [zz_generated_openapi.go](https://github.com/kubernetes/kubernetes/blob/master/pkg/generated/openapi/zz_generated.openapi.go) is a large map of all built-in resources. This map is regenerated as part of passing in [GetDefinitions](https://github.com/kubernetes/kubernetes/blob/master/staging/src/k8s.io/apiextensions-apiserver/pkg/controller/openapi/builder/builder.go#L540-L549) to the builder and then called when [building the OpenAPI](https://github.com/kubernetes/kube-openapi/blob/master/pkg/builder/openapi.go#L130-L133) for a particular CRD. 

This map is regenerated for every CRD resource, though discarded after the OpenAPI of a particular CRD is built so the memory is reclaimed. However, for large amounts of CRDs applied at once, this could cause a spike in memory usage.

The map is regenerated for every CRD because each CRD is appended to the map of all the built-ins and then regenerated. We can reuse the map by splitting the definitions into two maps and add a getter that checks the existence of a resource in two maps.

Benchmarks (benchmarking code has not been committed yet, but it spins up an apiserver, creates X CRDs, fetches the OpenAPI to ensure any lazy code is exercised and all CRD specs are successfully generated)

```
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/test/integration/apiserver/openapi
cpu: Intel(R) Xeon(R) CPU @ 2.20GHz
                                          │  old.txt   │             new.txt              │
                                          │   sec/op   │   sec/op    vs base              │
OpenAPIbyCRD/BenchmarkOpenAPIbyCRD-0-24     1.110 ± 0%   1.112 ± 1%       ~ (p=0.699 n=6)
OpenAPIbyCRD/BenchmarkOpenAPIbyCRD-10-24    2.429 ± 1%   2.455 ± 2%  +1.07% (p=0.015 n=6)
OpenAPIbyCRD/BenchmarkOpenAPIbyCRD-50-24    3.738 ± 1%   3.754 ± 1%       ~ (p=0.485 n=6)
OpenAPIbyCRD/BenchmarkOpenAPIbyCRD-150-24   6.494 ± 1%   6.519 ± 2%       ~ (p=0.937 n=6)
geomean                                     2.844        2.859       +0.52%

                                          │    old.txt    │               new.txt                │
                                          │     B/op      │     B/op       vs base               │
OpenAPIbyCRD/BenchmarkOpenAPIbyCRD-0-24      52.94Mi ± 1%   52.86Mi ±  1%        ~ (p=0.240 n=6)
OpenAPIbyCRD/BenchmarkOpenAPIbyCRD-10-24     203.1Mi ± 4%   197.2Mi ±  3%        ~ (p=0.132 n=6)
OpenAPIbyCRD/BenchmarkOpenAPIbyCRD-50-24     462.0Mi ± 8%   396.8Mi ±  1%  -14.12% (p=0.002 n=6)
OpenAPIbyCRD/BenchmarkOpenAPIbyCRD-150-24   1073.2Mi ± 7%   965.4Mi ± 10%  -10.04% (p=0.009 n=6)
geomean                                      270.2Mi        251.4Mi         -6.97%

                                          │   old.txt   │              new.txt               │
                                          │  allocs/op  │  allocs/op   vs base               │
OpenAPIbyCRD/BenchmarkOpenAPIbyCRD-0-24     139.2k ± 3%   139.1k ± 3%        ~ (p=0.394 n=6)
OpenAPIbyCRD/BenchmarkOpenAPIbyCRD-10-24    816.5k ± 1%   781.8k ± 1%   -4.24% (p=0.002 n=6)
OpenAPIbyCRD/BenchmarkOpenAPIbyCRD-50-24    2.224M ± 7%   1.890M ± 1%  -15.03% (p=0.002 n=6)
OpenAPIbyCRD/BenchmarkOpenAPIbyCRD-150-24   5.354M ± 5%   4.779M ± 8%  -10.73% (p=0.002 n=6)
geomean                                     1.078M        995.5k        -7.69%
```

#### Which issue(s) this PR fixes:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

#### Special notes for your reviewer:

#### Does this PR introduce a user-facing change?
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
NONE
```

#### Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.:

<!--
This section can be blank if this pull request does not require a release note.

When adding links which point to resources within git repositories, like
KEPs or supporting documentation, please reference a specific commit and avoid
linking directly to the master branch. This ensures that links reference a
specific point in time, rather than a document that may change over time.

See here for guidance on getting permanent links to files: https://help.github.com/en/articles/getting-permanent-links-to-files

Please use the following format for linking documentation:
- [KEP]: <link>
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs

```

/triage accepted