@alexzielenski 2023-01-18
ik this is WIP, but it would be helpful during review if you could split the openapi change to a separate leading commit