@logicalhan 2023-01-26
/remove-sig instrumentation