@Jefftree 2023-04-27
I'd like to get this in once we're able to properly measure the impact of these performance improvements. Definitely planned for the current release cycle 1.28.